% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding.R
\name{adding}
\alias{adding}
\title{Adds two numeric objects together}
\usage{
adding(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
numeric
}
\description{
Adds two numeric objects together
}
\examples{
adding(3,4)
}
