% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\title{Tracks genets through time}
\usage{
assign(dat, inv, dorm, buff, buffGenet, clonal, ...)
}
\description{
This function tracks individual plants through time, but only of
one species in one quadrat. It is designed for use within the
\code{\link{trackSpp}} function, and is not intended for use on its own.
}
\details{
see \code{\link{trackSpp}} for details of arguments and usage.
}
\seealso{
\code{\link[=trackSpp]{trackSpp()}}, which is a wrapper for the \code{\link[=assign]{assign()}} function that
applies it over many species and quadrats. The \code{\link[=assign]{assign()}} function uses the
\code{\link[=groupByGenet]{groupByGenet()}} function to group ramets into genets
(if 'clonal' argument = 1).
}
