% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawQuadMap.R
\name{drawQuadMap}
\alias{drawQuadMap}
\title{Create maps of quadrats over time, color-coded either by species or
by trackID}
\usage{
drawQuadMap(
  dat,
  type = "Species",
  species = "Species",
  site = "Site",
  quad = "Quad",
  year = "Year",
  geometry = "geometry",
  trackID = "trackID"
)
}
\arguments{
\item{trackID}{}
}
\value{

}
\description{
Create maps of quadrats over time, color-coded either by species or
by trackID
}
