% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasalAreas.R
\name{getBasalAreas}
\alias{getBasalAreas}
\title{Title}
\usage{
getBasalAreas(
  dat,
  species = "Species",
  quad = "Quad",
  site = "Site",
  year = "Year",
  geometry = "geometry"
)
}
\arguments{
\item{geometry}{}
}
\value{

}
\description{
Title
}
