% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighbors.R
\name{getNeighbors}
\alias{getNeighbors}
\title{getNeighbors}
\usage{
getNeighbors(
  dat,
  radius,
  method,
  compType = "allSpp",
  focal = "genet",
  trackID = "trackID",
  species = "Species",
  quad = "Quad",
  year = "Year",
  site = "Site",
  geometry = "geometry",
  ...
)
}
\arguments{
\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{

}
\description{

}
\details{

}
\examples{

}
