% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByGenet.R
\name{groupByGenet}
\alias{groupByGenet}
\title{This function will assign grouping variables for each feature of an sf
object. Grouping is based on buffers of a user-defined distance overlapping.}
\usage{
groupByGenet(sf, buffGenet)
}
\arguments{
\item{sf}{your sf object}

\item{buffGenet}{the distance you want the buffer to be (put 2.5 to group
polygons that are 5cm from each other)}
}
\value{

}
\description{
This function will assign grouping variables for each feature of an sf
object. Grouping is based on buffers of a user-defined distance overlapping.
}
