\name{plantTracker-package}
\alias{plantTracker-package}
\alias{plantTracker}
\docType{package}
\title{
\packageTitle{plantTracker}
}
\description{
\packageDescription{plantTracker}
}
\author{
\packageAuthor{plantTracker}

Maintainer: \packageMaintainer{plantTracker}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{plantTracker}
\packageIndices{plantTracker}
}

\seealso{
This vignette will guide you through using the plantTracker package
\code{\link[Suggested plantTracker Workflow.html]{Suggested_plantTracker_Workflow}}

This vignette goes through the trackSpp() function step-by-step
\code{\link[Using the plantTracker trackSpp function.html]{Using_the_plantTracker_trackSpp_function}}
}
