% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackSpp.R
\name{trackSpp}
\alias{trackSpp}
\title{Tracks genets through time for multiple species and sites}
\usage{
trackSpp(
  dat,
  inv,
  dorm = NULL,
  buff = NULL,
  buffGenet = NULL,
  clonal = NULL,
  ...
)
}
\value{

}
\description{
Tracks genets through time for multiple species and sites
}
