% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\title{This function tracks individual plants through time}
\usage{
assign(dat, inv, dorm, buff, buffGenet, clonal, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
This function tracks individual plants through time
}
\examples{

}
