% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasalAreas.R
\name{getBasalAreas}
\alias{getBasalAreas}
\title{Calculates basal area for each species in a quadrat}
\usage{
getBasalAreas(
  dat,
  species = "Species",
  quad = "Quad",
  site = "Site",
  year = "Year",
  geometry = "geometry"
)
}
\arguments{
\item{dat}{An sf data.frame in which each row represents a unique polygon
(either a genet or a ramet) in a unique site/quadrat/year combination. A
data.frame returned by \code{\link{trackSpp}} can be put directly into this
function. However, it is not necessary for 'dat' to have demographic data or
unique identifiers (i.e. 'trackID's) assigned. 'dat' must have columns that
contain a unique identification for each research site (default name is
"Site"), species name (default name is "Species"), quadrat identifier
(default name is "Quad"), year of data collection (default name is "Year"),
and an s.f 'geometry' column that contains a polygon or multipolygon data
type for each individual observation.}

\item{species}{An optional character string argument. Indicates
the name of the column in 'dat' that contains species name data. It is
unnecessary to include a value for this argument if the column name is
"Species" (default value is 'Species').}

\item{quad}{An optional character string argument. Indicates
the name of the column in 'dat' that contains quadrat name data. It is
unnecessary to include a value for this argument if the column name is
"Quad" (default is 'Quad').}

\item{site}{An optional character string argument. Indicates
the name of the column in 'dat' that contains site name data. It is
unnecessary to include a value for this argument if the column name is
"Site" (default value is 'Site').}

\item{year}{An optional character string argument. Indicates
the name of the column in 'dat' that contains data for year of sampling. It
is unnecessary to include a value for this argument if the column name is
"Year" (default is 'Year').}

\item{geometry}{An optional character string argument. Indicates
the name of the column in 'dat' that contains sf geometry data. It is
unnecessary to include a value for this argument if the column name is
"geometry" (default is 'geometry').}
}
\value{
\code{\link[=getBasalAreas]{getBasalAreas()}} returns a data.frame with the colums "Site",
"Quad", "Year" and "Species". No two rows will have the same values for all
four of these columns. The column "absolute_basalArea" has the area of the
quadrat that is occupied by a species in a given unique site/quadrat/year
combination. This is in the same units as the area for area in 'dat'.
"quad_basalArea" gives the combined basal area of all plants in a given
site/quadrat/year. "percent_basalArea" gives the percentage of occupied basal
area within a quadrat that is occupied by each species in a given
site/quadrat/year. For example, species A has a basal area of 22 cm^2 in
quadrat 1 in 2005 ("absolute_basalArea = 22). In 2005, there are 50 cm^2 of
quadrat 1 that are occupied by plants ("quad_basalArea" = 55). 44\% of the
occupied basal area in quadrat 1 in 2005 is occupied by species A
("percent_basalArea" = 44).
}
\description{
This function calculates the total basal area for each species
present in a quadrat for each year of data collection. Both the absolute
basal area (in the same units of area as the input data.frame) is returned,
as well as the percentage of the total occupied basal area in the quadrat
that is occupied by a given species ("percent total basal area").
}
