% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecruits.R
\name{getRecruits}
\alias{getRecruits}
\title{Calculates the number of recruits of each species per year in each quadrat}
\usage{
getRecruits(
  dat,
  species = "Species",
  quad = "Quad",
  site = "Site",
  year = "Year",
  trackID = "trackID",
  recruit = "recruit",
  byGenet = TRUE
)
}
\arguments{
\item{byGenet}{}
}
\value{

}
\description{
Calculates the number of recruits of each species per year in each quadrat
}
