% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasslandInventory.R
\docType{data}
\name{grasslandInventory}
\alias{grasslandInventory}
\title{A dataset indicating sampling years for the grasslandData example dataset}
\format{
A named list with 8 items, each of which if named for a quadrat
included in the grasslandData dataset.
\describe{
\item{C5}{An integer vector indicating the years in which quadrat C5 was
mapped. }
\item{D1}{An integer vector indicating the years in which quadrat D1 was
mapped.}
\item{SG2}{An integer vector indicating the years in which quadrat SG2
was mapped.}
\item{SG4}{An integer vector indicating the years in which quadrat SG4 was
mapped.}
...
}
}
\source{
\doi{dx.doi.org/10.1890/11-2200.1} \cr
\doi{doi.org/10.6084/m9.figshare.c.3304113.v1}
}
\usage{
grasslandInventory
}
\description{
This dataset indicates years in which sampling was conducted for each of the
four quadrats included in the \code{\link{grasslandData}} example dataset.
}
\keyword{datasets}
