\name{plantTracker-package}
\alias{plantTracker-package}
\alias{plantTracker}
\docType{package}
\title{
\packageTitle{plantTracker}
}
\description{
\packageDescription{plantTracker}
}
\author{
\packageAuthor{plantTracker}

Maintainer: \packageMaintainer{plantTracker}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{plantTracker}
\packageIndices{plantTracker}
}

\seealso{
Use the following code to see the vignette will guide you through using the plantTracker package
\code{vignette(topic = "Suggested_plantTracker_Workflow", package = "plantTracker")}

Use the following code to see the vignette goes through the trackSpp() function step-by-step
\code{vignette(topic = "Using_the_plantTracker_trackSpp_function", package = "plantTracker")}
}
