% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackSpp.R
\name{trackSpp}
\alias{trackSpp}
\title{Tracks mapped individual plants through time to get growth and survival data}
\usage{
trackSpp()
}
\value{

}
\description{
Tracks mapped individual plants through time to get growth and survival data
}
